modify 6-print_numberz.c

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -10,7 +10,7 @@ int main(void)
 
 	while (i < 'k')
 	{
-		putchar(i);
+		putchar((int)i);
 		i++;
 	}
 	putchar('\n');
